+++ inst/recode-regexp.sql
@@ -1,5 +1,5 @@
 select FILGRUPPE, LESID, KOL, TYPE, FRA, TIL
-from tbl_Kode
+from tbl_KodeBok
 where (FILGRUPPE = '%s' or FILGRUPPE = 'ALLE')
   and VERSJONTIL = #9999-01-01#
   and (LESID is NULL or LESID = '%s')